finish presence insert

+++ src/json.rs
@@ -61,7 +61,7 @@ impl Into<models::System> for System {
 #[serde(deny_unknown_fields)]
 pub struct MinorFactionPresence {
     pub state_id:Option<i32>,
-    pub influence:Option<f64>,
+    pub influence:Option<f32>,
     pub minor_faction_id:i32,
     pub state:Option<String>,
 }
